refactor some duplicated code into a function, general cleanup/tidying

+++ lib/tessa.pm
@@ -16,6 +16,12 @@ my %HTTP_STATUS_FOR = (
     SERVER_ERROR       => 500,
 );
 
+my %LENGTH_FOR = (
+   NAME => 128,
+   NOTE => 256,
+   URI  => 256, 
+);
+
 # / 
 # all HTTP verbs except for OPTIONS are not allowed for '/'
 any [qw(get post put del)] => '/' => \&_method_not_allowed;
@@ -46,10 +52,17 @@ options '/' => sub {
 get '/assets' => sub {
     my @assets;
 
-    my $assets = $db->get_all_assets();
-    if ( ! $assets ) {
-	return _throw_json_error('no assets exist', $HTTP_STATUS_FOR{NOT_FOUND});
-    }
+    my $assets;
+    eval { 
+	$assets	= $db->get_all_assets();
+    };
+    if ( my $err = $@ ) {
+	return _throw_json_error(
+  	   $HTTP_STATUS_FOR{SERVER_ERROR},
+	   "Error getting all asset records: '$err'", 
+	);
+    }	
+    
 
     return to_json { assets => $assets };
 };
@@ -59,15 +72,23 @@ post '/assets' => sub {
 
     if ( ! $request->{name} || ! $request->{uri} ) {
 	return _throw_json_error(
+	    $HTTP_STATUS_FOR{BAD_REQUEST},
 	    "Missing required parameter 'name' or 'uri' from JSON request body", 
-	    $HTTP_STATUS_FOR{BAD_REQUEST}
 	);
     }
-    
-    my $asset = $db->put_asset( 
-	$request->{name}, 
-	$request->{uri},
-    );
+  # TODO TEST FOR THIS  
+    _validate_asset_record( $request->{name}, $request->{uri} ); 
+	    
+    my $asset;
+    eval {
+	$asset = $db->put_asset( $request->{name}, $request->{uri} );
+    };
+    if ( my $err = $@ ) {
+   	return _throw_json_error( 
+	    $HTTP_STATUS_FOR{SERVER_ERROR},
+	    "Error creating new asset: '$err'", 
+	);
+    }
     return to_json($asset);     
 };
 
@@ -76,7 +97,10 @@ del '/assets' => sub {
 	$db->delete_all_assets;
     }; 
     if ( my $err = $@ ) {
-	return _throw_json_error("Error deleting all assets and notes: '$err'", $HTTP_STATUS_FOR{SERVER_ERROR});
+	return _throw_json_error(
+	    $HTTP_STATUS_FOR{SERVER_ERROR},
+	    "Error deleting all assets and notes: '$err'", 
+	);
     }
 
     return;
@@ -88,9 +112,16 @@ put '/assets' => \&_method_not_allowed;
 get '/assets/:asset_id' => sub {
     my $asset_id = route_parameters->get('asset_id');
 
-    my $asset = $db->get_asset( $asset_id );
-    if ( ! $asset ) {
-	return _throw_json_error("asset '$asset_id' does not exist", $HTTP_STATUS_FOR{NOT_FOUND});
+    my $asset;
+    eval {
+	return unless _asset_exists( $asset_id );
+        $asset = $db->get_asset( $asset_id );
+    };
+    if ( my $err = $@ ) {
+	return _throw_json_error(
+	    $HTTP_STATUS_FOR{SERVER_ERROR},
+            "Error retrieving asset '$asset_id': $err", 
+	);
     }
     
     return to_json $asset;
@@ -102,18 +133,24 @@ put '/assets/:asset_id' => sub {
 
     if ( ! $request->{name} && ! $request->{uri} ) {
 	return _throw_json_error(
+	    $HTTP_STATUS_FOR{BAD_REQUEST},
 	    "Must supply 'name' or 'uri' in JSON request body to update an asset", 
-	    $HTTP_STATUS_FOR{BAD_REQUEST}
 	);
     }
     
-    my $asset = $db->update_asset( 
-	$asset_id,
-	$request->{name}, 
-	$request->{uri},
-    );
-    if ( ! $asset ) {
-	return _throw_json_error("asset '$asset_id' does not exist", $HTTP_STATUS_FOR{NOT_FOUND});
+    # TODO TEST FOR THIS  
+    _validate_asset_record( $request->{name}, $request->{uri} ); 
+
+    my $asset;
+    eval {
+	return unless _asset_exists( $asset_id );
+	$asset = $db->update_asset( $asset_id, $request->{name}, $request->{uri} );
+    };
+    if ( my $err = $@ ) {
+	return _throw_json_error(
+	    $HTTP_STATUS_FOR{SERVER_ERROR},
+            "Error updating asset '$asset_id': $err", 
+        );
     }
 
     return to_json $asset;
@@ -126,7 +163,10 @@ del '/assets/:asset_id' => sub {
 	$db->delete_asset( $asset_id );
     }; 
     if ( my $err = $@ ) {
-	return _throw_json_error("Error deleting asset '$asset_id': $err", $HTTP_STATUS_FOR{SERVER_ERROR});
+	return _throw_json_error(
+	    $HTTP_STATUS_FOR{SERVER_ERROR},
+            "Error deleting asset '$asset_id': $err", 
+        );
     }
     
     return;
@@ -141,14 +181,13 @@ get '/assets/:asset_id/notes' => sub {
 
     my $notes;
     eval {
-	_asset_exists( $asset_id ) 
-	    or return _throw_json_error("asset '$asset_id' does not exist", $HTTP_STATUS_FOR{NOT_FOUND});
+	return unless _asset_exists( $asset_id );
 	$notes = $db->get_all_notes_for_asset( $asset_id );
     };
     if ( my $err = $@ ) {
 	return _throw_json_error(
+	    $HTTP_STATUS_FOR{SERVER_ERROR},
 	    "Error getting notes for asset '$asset_id': $err", 
-	    $HTTP_STATUS_FOR{SERVER_ERROR}
 	);
     }
 
@@ -161,21 +200,23 @@ post '/assets/:asset_id/notes' => sub {
     my $request = from_json request->body;
     if ( ! $request->{note} ) {
 	return _throw_json_error(
+	    $HTTP_STATUS_FOR{BAD_REQUEST},
 	    "Missing required parameter 'note' in JSON request body", 
-	    $HTTP_STATUS_FOR{BAD_REQUEST}
 	);
     }
-    
+   
+    # TODO TEST FOR THIS 
+    _validate_note_record( $request->{note} );
+ 
     my $asset;
     eval {
-	_asset_exists( $asset_id ) 
-	    or return _throw_json_error("asset '$asset_id' does not exist", $HTTP_STATUS_FOR{NOT_FOUND});
+	return unless _asset_exists( $asset_id );
 	$asset = $db->put_note_for_asset( $asset_id, $request->{note});
     };
     if ( my $err = $@ ) {
 	return _throw_json_error(
+	    $HTTP_STATUS_FOR{SERVER_ERROR},
 	    "Error creating new note record for asset '$asset_id': $err", 
-	    $HTTP_STATUS_FOR{SERVER_ERROR}
 	);
     }
 
@@ -190,8 +231,8 @@ del '/assets/:asset_id/notes' => sub {
     }; 
     if ( my $err = $@ ) {
 	return _throw_json_error(
+	    $HTTP_STATUS_FOR{SERVER_ERROR},
 	    "Error deleting all note records for asset '$asset_id': $err", 
-	    $HTTP_STATUS_FOR{SERVER_ERROR}
 	);
     }
     
@@ -209,17 +250,17 @@ put '/assets/:asset_id/notes/:note_id' => sub {
     my $request = from_json request->body;
     if ( ! $request->{note} ) {
 	return _throw_json_error(
+	    $HTTP_STATUS_FOR{BAD_REQUEST},
 	    "Missing required parameter 'note' in JSON request body", 
-	    $HTTP_STATUS_FOR{BAD_REQUEST}
 	);
     }
 
+    _validate_note_record( $request->{note} );
+
     my $asset;
     eval {
-	_asset_exists( $asset_id ) 
-	    or return _throw_json_error("asset '$asset_id' does not exist", $HTTP_STATUS_FOR{NOT_FOUND});
-	_asset_owns_note( $asset_id, $note_id )
-	    or return _throw_json_error("asset '$asset_id' does not own note '$note_id'", $HTTP_STATUS_FOR{BAD_REQUEST});
+	return unless _asset_exists( $asset_id );
+	return unless _asset_owns_note( $asset_id, $note_id );
 	$asset = $db->update_note_for_asset( $asset_id, $note_id, $request->{note} );
     };
 
@@ -231,16 +272,13 @@ del '/assets/:asset_id/notes/:note_id' => sub {
     my $note_id  = route_parameters->get('note_id');
 
     eval {
-	_asset_exists( $asset_id ) 
-	    or return _throw_json_error("asset '$asset_id' does not exist", $HTTP_STATUS_FOR{NOT_FOUND});
-	_asset_owns_note( $asset_id, $note_id )
-	    or return _throw_json_error("asset '$asset_id' does not own note '$note_id'", $HTTP_STATUS_FOR{BAD_REQUEST});
+	return unless _asset_owns_note( $asset_id, $note_id );
 	$db->delete_note_for_asset( $asset_id, $note_id );
     }; 
     if ( my $err = $@ ) {
 	return _throw_json_error(
+	    $HTTP_STATUS_FOR{SERVER_ERROR},
 	    "Error deleting note record '$note_id' for asset '$asset_id': $err", 
-	    $HTTP_STATUS_FOR{SERVER_ERROR}
 	);
     }
 
@@ -249,40 +287,82 @@ del '/assets/:asset_id/notes/:note_id' => sub {
 
 any [qw(post get)] => '/assets/:asset_id/notes/:note_id' => \&_method_not_allowed;
 
+sub _validate_asset_record {
+    my ($name, $uri) = @_;
+	
+    my @errors; 
+    push @errors, "name '$name' is > $LENGTH_FOR{NAME} characters"
+	if $name && length $name > $LENGTH_FOR{NAME};
+    push @errors, "uri '$uri' is > $LENGTH_FOR{URI} characters"
+	if $uri && length $uri > $LENGTH_FOR{URI};
+
+    return _throw_json_error(
+	$HTTP_STATUS_FOR{BAD_REQUEST},
+	@errors,
+    ) if @errors;
+    return 1;
+}
+
+sub _validate_note_record {
+    my ($note) = @_;
+	
+    my @errors; 
+    push @errors, "note '$note' is > $LENGTH_FOR{NOTE} characters"
+	if length $note > $LENGTH_FOR{NOTE};
+
+    return _throw_json_error(
+	$HTTP_STATUS_FOR{BAD_REQUEST},
+	@errors,
+    ) if @errors;
+    return 1;
+}
+
 sub _asset_exists {
     my ($asset_id) = @_;
 
-    return $db->get_asset( $asset_id ) ? 1 : 0;
+    my $asset = $db->get_asset( $asset_id );
+    return _throw_json_error(
+        $HTTP_STATUS_FOR{NOT_FOUND},
+	"asset '$asset_id' does not exist", 
+    ) unless $asset;
+
+    return 1;
 }
 
 sub _asset_owns_note {
     my ($asset_id, $note_id) = @_;
 
-    return grep { 
+    return _throw_json_error( 
+	$HTTP_STATUS_FOR{BAD_REQUEST},
+	"asset '$asset_id' does not own note '$note_id'",
+    ) unless grep { 
 	$_->{id} == $note_id 
     } @{ $db->get_asset( $asset_id )->{notes} };
 }
 
 sub _build_error_json {
-    my ($error_message) = @_;
+    my (@error_messages) = @_;
 
-    return to_json { errors => [ $error_message ] };
+    return to_json { errors => [ @error_messages ] };
 }
 
 sub _throw_json_error {
-    my ($error_message, $http_status_code) = @_;
+    my ($http_status_code, @error_messages) = @_;
 
     Dancer2::Core::Error->new(
 	response     => response(),
 	status       => $http_status_code,
-	content      => _build_error_json($error_message),
+	content      => _build_error_json(@error_messages),
 	content_type => 'application/json',
     )->throw;    
     return;
 }
 
 sub _method_not_allowed {
-    return _throw_json_error('METHOD NOT ALLOWED', $HTTP_STATUS_FOR{METHOD_NOT_ALLOWED});
+    return _throw_json_error(
+	$HTTP_STATUS_FOR{METHOD_NOT_ALLOWED},
+	'METHOD NOT ALLOWED',
+    );
 }
 
 true;
